lastOf

+++ src/app/shared/info.service.ts
@@ -15,10 +15,15 @@ export class InfoService{
     readonly albums_root='https://jsonplaceholder.typicode.com/albums';
     
     public deletedImageId:number[]=[];
+    public updatedImages:Image[]=[];
+    public uploadedImages:Image[]=[];
+
 
 
    
     constructor(private http:HttpClient){}
+
+
 
     getImages():Observable<Image[]>{
      
@@ -27,14 +32,66 @@ export class InfoService{
         );
 
     }
+    
+    getAlbums():Observable<Album[]>{
+     
+      return  this.http.get<Album[]>(this.albums_root).pipe(
+           catchError(this.handleError)
+       );
+
+   }
     deleteImage(id){
+
       this.deletedImageId.push(id);
-      this.http.delete<Image>(this.photos+`/${id}`).pipe(catchError(this.handleError)).subscribe(response=>console.log(response));
+      if(id<=5000){
+        this.http.delete<Image>(this.photos+`/${id}`).pipe(catchError(this.handleError)).subscribe(response=>console.log(response));
+      }else{
+        this.uploadedImages.forEach(img=>{
+          if(img.id===id){
+          delete this.uploadedImages[this.uploadedImages.indexOf(img)];
+          }
+        })
+      }
     }
-    updateImage(id,body){
-      console.log(id)
-      console.log(body)
-      return this.http.put<Image>(this.photos+`/${id}`,body).pipe(catchError(this.handleError));
+    uploadImage(body){
+      this.http.post<Image>(this.photos,body).pipe(catchError(this.handleError)).subscribe(image=>{
+        console.log(image)
+        this.uploadedImages.push(image);
+      })
+  
+    }
+    updateImage(id,body:Image){
+      if(id>5000){
+        console.log('5000')
+        let editedImage:Image;
+        this.updatedImages.forEach(img=>{
+             if(img.id===body.id){
+               editedImage=img;
+             }
+        })
+        if(this.updatedImages.includes(editedImage)){
+          this.updatedImages[this.updatedImages.indexOf(editedImage)]=body;
+        }else{
+          console.log(body);
+          this.updatedImages.push(body);
+        }
+      }
+     else{
+      this.http.put<Image>(this.photos+`/${id}`,body).pipe(catchError(this.handleError)).subscribe(image=>{
+       
+        let editedImage:Image;
+        this.updatedImages.forEach(img=>{
+             if(img.id===image.id){
+               editedImage=img;
+             }
+        })
+        if(this.updatedImages.includes(editedImage)){
+          this.updatedImages[this.updatedImages.indexOf(editedImage)]=image;
+        }else{
+          this.updatedImages.push(image);
+        }
+      });
+    }
     }
   
     getImageById(id):Observable<Image>{
